+++ global-client/src/pages/App.vue
@@ -10,8 +10,8 @@
 					<Modal :show="global.isModalVisible">
 						<div class="flex flex-col justify-between h-full">
 							<div class="flex-1 text-center">
-								<router-link to="/" v-slot="{ isActive }">
-									<HubIcon type="pubhubs-home" :active="isActive" class="text-white" @click="toggleMenu.toggleMenu()"></HubIcon>
+								<router-link to="/">
+									<Icon type="pubhubs-home" size="3xl" class="text-white mx-auto" @click="toggleMenu.toggleMenu()"></Icon>
 								</router-link>
 
 								<Line v-if="global.loggedIn && global.hasPinnedHubs" class="m-6 mt-8"></Line>
@@ -23,8 +23,8 @@
 								<div v-if="global.loggedIn">
 									<SettingsDialog v-if="settingsDialog" @close="settingsDialog = false"></SettingsDialog>
 									<div class="flex justify-between">
-										<HubIcon type="cog" class="text-dark" size="lg" @click="settingsDialog = true"></HubIcon>
-										<HubIcon type="power" class="text-dark" size="lg" @click="logout()"></HubIcon>
+										<Icon type="cog" class="text-dark" size="lg" @click="settingsDialog = true"></Icon>
+										<Icon type="power" class="text-dark" size="lg" @click="logout()"></Icon>
 									</div>
 								</div>