+++ global-client/src/store/global.ts
@@ -91,19 +91,28 @@ const useGlobal = defineStore('global', {
 		setGlobalSettings(data: any) {
 			const settings = useSettings();
 			settings.setTheme(data.theme);
-			if (!data.timeformat || data.timeformat == '') {
+			if (!data.timeformat || data.timeformat === '') {
 				data.timeformat = TimeFormat.format24;
 			}
 			settings.setTimeFormat(data.timeformat);
-			if (!data.language || data.language == '') {
+			if (!data.language || data.language === '') {
 				data.language = navigator.language;
 			}
 			settings.setLanguage(data.language);
 			this.pinnedHubs = data.hubs;
 		},
 
-		login() {
-			window.location.replace(api.apiURLS.login);
+		login(language: string | 'en' | 'nl') {
+			switch (language) {
+				case 'en':
+					window.location.assign(api.apiURLS.loginEn);
+					break;
+				case 'nl':
+					window.location.assign(api.apiURLS.login);
+					break;
+				default:
+					window.location.assign(api.apiURLS.login);
+			}
 		},
 
 		logout() {
@@ -146,7 +155,7 @@ const useGlobal = defineStore('global', {
 
 		existsInPinnedHubs(hubId: string) {
 			if (!this.pinnedHubs) return false;
-			const found = this.pinnedHubs.find((hub) => hub.hubId == hubId);
+			const found = this.pinnedHubs.find((hub) => hub.hubId === hubId);
 			return found;
 		},