+++ hub-client/src/assets/icons.ts
@@ -23,7 +23,7 @@ const icons: { [key: string]: string } = {
         <circle cx="12" cy="12" r="9" fill="currentColor" />
         `,
 	'pubhubs-home': `
-        <circle cx="12" cy="12" r="9" fill="currentColor"/>
+        <circle cx="12" cy="12" r="11" fill="currentColor"/>
         <path d="M4.25306 12.2951L17.0548 0.626046L13.3499 17.5472L0.548216 29.2163L4.25306 12.2951Z" fill="black" stroke="black" stroke-linejoin="round" transform="scale(0.5) translate(15,9)"/>
         <circle cx="12" cy="12" r="1" fill="currentColor"/>
         `,
@@ -317,6 +317,15 @@ const icons: { [key: string]: string } = {
         </defs>
         </svg>
         `,
+	message: `
+        <svg class="h-32 w-32 text-gray-700"  width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">  <path stroke="none" d="M0 0h24v24H0z"/>  <path d="M21 14l-3 -3h-7a1 1 0 0 1 -1 -1v-6a1 1 0 0 1 1 -1h9a1 1 0 0 1 1 1v10" />  <path d="M14 15v2a1 1 0 0 1 -1 1h-7l-3 3v-10a1 1 0 0 1 1 -1h2" /></svg>`,
+	mention: `
+        <svg class="h-32 w-32 text-gray-700"  fill="none" viewBox="0 0 24 24" stroke="currentColor">
+  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M16 12a4 4 0 10-8 0 4 4 0 008 0zm0 0v1.5a2.5 2.5 0 005 0V12a9 9 0 10-9 9m4.5-1.206a8.959 8.959 0 01-4.5 1.207"/>
+</svg>`,
+	hub_fallback: `
+  <path stroke-linecap="round" stroke-linejoin="round" d="M11.412 15.655 9.75 21.75l3.745-4.012M9.257 13.5H3.75l2.659-2.849m2.048-2.194L14.25 2.25 12 10.5h8.25l-4.707 5.043M8.457 8.457 3 3m5.457 5.457 7.086 7.086m0 0L21 21" />
+`,
 };
 
 export { icons, iconSizes as sizes };