+++ hub-client/src/components/rooms/MessageSnippet.vue
@@ -1,10 +1,10 @@
 <template>
-	<div class="bg-hub-background-3 flex items-center rounded-md w-6/6 cursor-pointer">
-		<p v-if="showInReplyTo" class="ml-1 mr-1 shrink-0">{{ $t('message.in_reply_to') }}</p>
-		<H3 class="mr-2 ml-2 my-0 shrink-0" :class="textColor(userColor)">
+	<div class="bg-hub-background-3 flex px-2 gap-3 items-center rounded-md cursor-pointer">
+		<p v-if="showInReplyTo" class="text-nowrap">{{ $t('message.in_reply_to') }}</p>
+		<p :class="textColor(userColor)">
 			<UserDisplayName :user="event.sender"></UserDisplayName>
-		</H3>
-		<p class="mr-2 truncate">{{ text }}</p>
+		</p>
+		<p class="truncate" :title="text">{{ text }}</p>
 	</div>
 </template>