+++ hub-client/src/components/rooms/RoomList.vue
@@ -1,18 +1,22 @@
 <template>
 	<Menu>
-		<InlineSpinner v-if="!rooms.hasRooms"></InlineSpinner>
+		<InlineSpinner v-if="!rooms.roomsLoaded"></InlineSpinner>
 		<!-- There is always the general room -->
 		<template v-if="rooms.hasRooms">
 			<template v-for="room in rooms.sortedRoomsArray" :key="room.roomId">
 				<div v-if="showRoom(room)" :key="room.roomId" class="group" @click="toggleMenu.toggleGlobalMenu()">
-					<Icon type="unlink" class="cursor-pointer hover:text-red ml-2 float-right hidden group-hover:block" @click="leaveRoom(room.roomId)"></Icon>
 					<router-link :to="{ name: 'room', params: { id: room.roomId } }" v-slot="{ isActive }">
-						<Badge v-if="room.numUnreadMessages > 0" class="-ml-1 -mt-1">{{ room.numUnreadMessages }}</Badge>
-						<MenuItem :roomInfo="room" :icon="roomIcon(room)" :active="isActive">
-							<PrivateRoomName v-if="room.isPrivateRoom()" :members="room.getOtherMembers()"></PrivateRoomName>
-							<span v-else>
-								{{ room.name }}
-							</span>
+						<MenuItem :roomInfo="room" :icon="roomIcon(room)" :active="isActive" :key="room.roomId" class="relative">
+							<PrivateRoomName v-if="room.isPrivateRoom()" :members="room.getOtherJoinedAndInvitedMembers()"></PrivateRoomName>
+							<RoomName v-else :room="room"></RoomName>
+
+							<div class="absolute top-2 right-14">
+								<Icon type="unlink" class="fixed cursor-pointer invisible hover:text-red group-hover:visible" @click.prevent="leaveRoom(room.roomId)"></Icon>
+								<div v-if="settings.isFeatureEnabled(featureFlagType.notifications)" class="fixed visible group-hover:invisible">
+									<UnreadMessageBadge v-if="room.getRoomUnreadNotificationCount(NotificationCountType.Total) > 0">{{ room.getRoomUnreadNotificationCount(NotificationCountType.Total) }}</UnreadMessageBadge>
+									<UnreadMentionBadge v-if="room.getRoomUnreadNotificationCount(NotificationCountType.Highlight) > 0">{{ room.getRoomUnreadNotificationCount(NotificationCountType.Highlight) }}</UnreadMentionBadge>
+								</div>
+							</div>
 						</MenuItem>
 					</router-link>
 				</div>
@@ -29,6 +33,9 @@
 	import { RoomType } from '@/store/rooms';
 	import { usePlugins, PluginProperties } from '@/store/plugins';
 	import { useToggleMenu } from '@/store/toggleGlobalMenu';
+	import { NotificationCountType } from 'matrix-js-sdk';
+	import { useSettings, featureFlagType } from '@/store/store';
+	const settings = useSettings();
 
 	const { t } = useI18n();
 	const router = useRouter();
@@ -50,10 +57,10 @@
 		}
 		if (props.roomType !== '') {
 			const type = props.roomType.substring(1);
-			if (props.roomType.charAt(0) == '!') {
+			if (props.roomType.charAt(0) === '!') {
 				return room.getType() !== type;
 			} else {
-				return room.getType() == props.roomType;
+				return room.getType() === props.roomType;
 			}
 		}
 		return true;