+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Changes to main, not yet committed to stable
 
 *Please add a brief description of any changes to be tested and any migrations to be performed here.*
+- Some changes to styling of displayed hubs in global client. 
+
+## 9 May 2024
+- Add notifications and mentions in a better way.
 
 ## 30 Apr 2024
 - Refactured a lot for performance issues, pleas check the 'Restricted' room on stable. Does it has errors in the console? (see #634,#590,#636)